requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.103.2
ffmpy==0.3.2
filelock==3.12.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.34.0
gradio_client==0.2.6
h11==0.14.0
hi-ml-multimodal==0.1.2
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.5
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openpyxl==3.1.2
orjson==3.9.7
packaging==24.0
pandas==1.3.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.5.3
pydantic_core==2.14.6
pydicom==2.2.2
pydub==0.25.1
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.2
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
python-multipart==0.0.8
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sacremoses==0.0.53
scikit-image==0.18.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
semantic-version==2.10.0
SimpleITK==2.1.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
starlette==0.27.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
toolz==0.12.1
torch==1.13.0
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.2
transformers==4.17.0
typing_extensions==4.7.1
uc-micro-py==1.0.3
urllib3==2.0.7
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.4
zipp==3.15.0
